requirements.txt
Django==3.2.6
gunicorn==20.1.0
whitenoise==5.3.0
psycopg2-binary==2.9.1
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
